scraper/requirements.txt
openai
python-dotenv
requests
feedparser
beautifulsoup4
newspaper3k
